+++ test/dummy/test/integration/accept_terms_test.rb
@@ -36,7 +36,7 @@ class AcceptTermsTest < ActionDispatch::IntegrationTest
     assert_includes response.body, published_on
     refute_includes response.body, "ago"
     refute_includes response.body, "Read the full terms"
-    assert_select "div.py-5"
+    assert_select "div.pt-2"
     assert_select "input[type=checkbox][name=agreed][required]"
     assert_select "input[type=checkbox][name=agreed][checked]", count: 1
     assert_includes response.body, "Agree"
@@ -70,12 +70,13 @@ class AcceptTermsTest < ActionDispatch::IntegrationTest
     assert_response :success
     assert RecordingStudioTermsAndConditions.reaccepting?(@user, @workspace)
     refute RecordingStudioTermsAndConditions.accepted?(@user, @workspace)
-    assert_includes response.body, "Terms updated"
+    refute_includes response.body, "Terms updated"
     published_on = revised.current_publishable.publish_at.in_time_zone.strftime("%e %b %Y").squish
     assert_includes response.body, published_on
-    assert_includes CGI.unescapeHTML(response.body), "You already agreed. This version is from #{published_on}."
+    refute_includes CGI.unescapeHTML(response.body), "You already agreed. This version is from #{published_on}."
     refute_includes response.body, "What changed"
     assert_select "button[type=submit]", text: "Agree again"
+    assert_select "div.pt-2"
     refute_includes CGI.unescapeHTML(response.body), "The live version for this workspace."
     refute_includes CGI.unescapeHTML(response.body), "These terms changed. Agree again to stay in."
   end